+++ jest.config.js
@@ -1,4 +1,11 @@
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node'
+  preset: "ts-jest",
+  testEnvironment: "node",
+
+  globals: {
+    "ts-jest": {
+      // Disable type checking in test
+      diagnostics: false,
+    },
+  },
 };