+++ src/components/parser.ts
@@ -1,5 +1,6 @@
 import * as fs from "fs";
 import * as ts from "typescript";
+import { defaultConfig, IConfigurator, IInterfaceSorterConfiguration, SimpleConfigurator } from "./configurator";
 
 export interface IComment {
   text: string;
@@ -18,28 +19,45 @@ export interface ITsNode {
   comments: IComments;
 }
 
-export interface ITsInterfaceNode extends ITsNode {
-  declaration: ts.InterfaceDeclaration;
-  members: Array<ITsInterfaceMemberNode>;
+export interface ITsNodeWithMembers extends ITsNode {
+  members: Array<ITsTypeMemberNode>;
 }
 
-export interface ITsInterfaceMemberNode extends ITsNode {
+export interface ITsInterfaceNode extends ITsNodeWithMembers {
+  declaration: ts.InterfaceDeclaration;
+}
+
+export interface ITsTypeLiteralNode extends ITsNodeWithMembers {
+  declaration: ts.TypeAliasDeclaration;
+}
+
+export type ITypeNode = ITsInterfaceNode | ITsTypeLiteralNode;
+
+export interface ITsTypeMemberNode extends ITsNode {
   element: ts.TypeElement;
   text: string;
 }
 
 export interface ITsParser {
-  parseInterface(
+  parseTypeNodes(
     fullFilePath: string,
     sourceFileText?: string
-  ): { nodes: ITsInterfaceNode[]; sourceFile: ts.SourceFile | null };
+  ): { nodes: ITypeNode[]; sourceFile: ts.SourceFile | null };
 }
 
 export class SimpleTsParser implements ITsParser {
-  public parseInterface(
+  private configurator: IConfigurator<IInterfaceSorterConfiguration>;
+
+  public constructor(
+    configurator: IConfigurator<IInterfaceSorterConfiguration>
+  ) {
+    this.configurator = configurator;
+  }
+
+  public parseTypeNodes(
     fullFilePath: string,
     _sourceText?: string | undefined
-  ): { nodes: ITsInterfaceNode[]; sourceFile: ts.SourceFile | null } {
+  ): { nodes: ITypeNode[]; sourceFile: ts.SourceFile | null } {
     if (_sourceText !== null && _sourceText !== undefined && _sourceText.trim() === "") {
       return { nodes: [], sourceFile: null };
     }
@@ -58,21 +76,21 @@ export class SimpleTsParser implements ITsParser {
   private delintFile(
     sourceFile: ts.SourceFile,
     sourceText?: string
-  ): { interfaces: ITsInterfaceNode[] } {
+  ): { interfaces: ITypeNode[] } {
     const sourceFileText = sourceText || sourceFile.getText();
 
-    return { interfaces: this.delintInterfaces(sourceFile, sourceFile, sourceFileText) };
+    return { interfaces: this.delintTypeNodes(sourceFile, sourceFile, sourceFileText) };
   }
 
-  private delintInterfaces(
+  private delintTypeNodes(
     node: ts.Node,
     sourceFile: ts.SourceFile,
     sourceFileText: string
-  ): ITsInterfaceNode[] {
-    const interfaceNodes: ITsInterfaceNode[] = [];
+  ): ITypeNode[] {
+    const typeNodes: ITypeNode[] = [];
 
     switch (node.kind) {
-      case ts.SyntaxKind.InterfaceDeclaration:
+      case ts.SyntaxKind.InterfaceDeclaration: {
         const interfaceNode = node as ts.InterfaceDeclaration;
         const lines = this.getCodeLineNumbers(node, sourceFile);
         const delintedInterface: ITsInterfaceNode = {
@@ -80,19 +98,42 @@ export class SimpleTsParser implements ITsParser {
           start: lines.start,
           end: lines.end,
           comments: this.getComments(node, sourceFileText),
-          members: this.delintInterfaceElements(interfaceNode, sourceFile, sourceFileText)
+          members: this.getMembers(interfaceNode, sourceFile, sourceFileText)
         };
-        interfaceNodes.push(delintedInterface);
+        typeNodes.push(delintedInterface);
         break;
+      }
+      case ts.SyntaxKind.TypeAliasDeclaration: {
+        const sortTypes = this.configurator.getValue("sortTypes") as boolean;
+        // Only parse `type` when settings is true
+        if (!sortTypes) {
+          break;
+        }
+
+        const typeNode = node as ts.TypeAliasDeclaration;
+        const lines = this.getCodeLineNumbers(node, sourceFile);
+        if (typeNode.type.kind === ts.SyntaxKind.TypeLiteral) {
+          const members = this.getMembers(typeNode.type as ts.TypeLiteralNode, sourceFile, sourceFileText);
+          const delintedInterface: ITsTypeLiteralNode = {
+            declaration: typeNode,
+            start: lines.start,
+            end: lines.end,
+            comments: this.getComments(node, sourceFileText),
+            members
+          };
+          typeNodes.push(delintedInterface);
+        }
+        break;
+      }
       default:
         break;
     }
 
     ts.forEachChild(node, node => {
-      interfaceNodes.push(...this.delintInterfaces(node, sourceFile, sourceFileText));
+      typeNodes.push(...this.delintTypeNodes(node, sourceFile, sourceFileText));
     });
 
-    return interfaceNodes;
+    return typeNodes;
   }
 
   private getComments(node: ts.Node, sourceFileText: string): IComments {
@@ -121,11 +162,11 @@ export class SimpleTsParser implements ITsParser {
     return { start, end };
   }
 
-  private delintInterfaceElements(
-    node: ts.InterfaceDeclaration,
-    sourceFile: ts.SourceFile,
-    sourceFileText: string
-  ): ITsInterfaceMemberNode[] {
+  private getMembers<T extends { readonly members: ts.NodeArray<ts.TypeElement>; }>
+    (
+      node: T,
+      sourceFile: ts.SourceFile,
+      sourceFileText: string) {
     return node.members.map(x => {
       const lines = this.getCodeLineNumbers(x, sourceFile);
       const comments = this.getComments(x, sourceFileText);