+++ src/test/components/sorter.test.ts
@@ -1,32 +1,27 @@
 import { SimpleTsParser } from "../../components/parser";
 import { SimpleTsSorter } from "../../components/sorter";
 import {
+  defaultConfig,
   IInterfaceSorterConfiguration,
   SimpleConfigurator,
 } from "../../components/configurator";
 
 import {
+  tcPrefixes,
   tcClassImplementInterface,
   tcEmptyInterface,
   tcPrefixExport,
-  tcInterfaceWithOneProperty,
   tcInterfaceWithExtends,
-  tcInterfaceWithComment,
-  tcInterfaceWithJsDocProperty,
   tcInterfaceWithOptionalProperty,
   tcInterfaceWithCapitalLetter,
   tcInterfaceWithMultipleOptionalProperty,
+  typeWithJsDocProperty,
+  typeWithOneProperty,
 } from "./test-cases";
 
 describe("Sorter", () => {
-  const defaultConfig: IInterfaceSorterConfiguration = {
-    lineBetweenMembers: true,
-    indentSpace: 2,
-    sortByCapitalLetterFirst: false,
-    sortByRequiredElementFirst: false,
-  };
 
-  const parser = new SimpleTsParser();
+  const parser = new SimpleTsParser(new SimpleConfigurator({ default: defaultConfig }));
   const sorter = new SimpleTsSorter(
     new SimpleConfigurator({ default: defaultConfig })
   );
@@ -34,37 +29,40 @@ describe("Sorter", () => {
   const filePath = "Untitled-1";
 
   test("should not break with no interface", () => {
-    const { nodes } = parser.parseInterface(
+    const { nodes } = parser.parseTypeNodes(
       filePath,
       tcClassImplementInterface
     );
-    const sorted = sorter.sortInterfaceElements(nodes);
+    const sorted = sorter.sortGenericTypeElements(nodes);
 
     expect(sorted.length).toBe(0);
   });
 
   test("should not sort with empty interface", () => {
-    const { nodes } = parser.parseInterface(filePath, tcEmptyInterface);
-    const sorted = sorter.sortInterfaceElements(nodes);
+    const { nodes } = parser.parseTypeNodes(filePath, tcEmptyInterface);
+    const sorted = sorter.sortGenericTypeElements(nodes);
 
     expect(sorted.length).toBe(0);
   });
 
-  test("should sort one interface with export prefix", () => {
-    const { nodes } = parser.parseInterface(
+  test.each([tcPrefixes])("should sort one with export prefix - %s", (prefix) => {
+    const prefixLength = prefix.length;
+    const exportLength = tcPrefixExport.length;
+    const { nodes } = parser.parseTypeNodes(
       filePath,
-      tcPrefixExport + tcInterfaceWithOneProperty
+      tcPrefixExport + prefix + typeWithOneProperty
     );
-    const sorted = sorter.sortInterfaceElements(nodes);
+    const sorted = sorter.sortGenericTypeElements(nodes);
 
     expect(sorted.length).toBe(1);
-    expect(sorted[0].rangeToRemove).toEqual({ pos: 30, end: 46 });
+    //  
+    expect(sorted[0].rangeToRemove).toEqual({ pos: prefixLength + exportLength + 1, end: prefixLength + exportLength + 17 });
   });
 
   test("should sort two interfaces with one extends the other", () => {
     const textIntput = tcInterfaceWithExtends;
-    const { nodes } = parser.parseInterface(filePath, textIntput);
-    const sorted = sorter.sortInterfaceElements(nodes);
+    const { nodes } = parser.parseTypeNodes(filePath, textIntput);
+    const sorted = sorter.sortGenericTypeElements(nodes);
 
     expect(sorted.length).toBe(1);
     expect(sorted[0].rangeToRemove).toEqual({ pos: 75, end: 110 });
@@ -73,15 +71,16 @@ describe("Sorter", () => {
     );
   });
 
-  test("should sort one interface with comments", () => {
-    const textIntput = tcInterfaceWithJsDocProperty;
-    const { nodes } = parser.parseInterface(filePath, textIntput);
-    const sorted = sorter.sortInterfaceElements(nodes);
+  test.each(tcPrefixes)(`should sort one type with comments - %s`, (prefix) => {
+    const prefixLength = prefix.length;
+    const textIntput = prefix + typeWithJsDocProperty;
+    const { nodes } = parser.parseTypeNodes(filePath, textIntput);
+    const sorted = sorter.sortGenericTypeElements(nodes);
 
     expect(sorted.length).toBe(1);
-    expect(sorted[0].rangeToRemove).toEqual({ pos: 23, end: 143 });
+    expect(sorted[0].rangeToRemove).toEqual({ pos: prefixLength + 1, end: prefixLength + 121 });
     expect(sorted[0].textToReplace).toEqual(
-      textIntput.substring(93, 143) + "\n" + textIntput.substring(23, 92)
+      textIntput.substring(prefixLength + 71, prefixLength + 121) + "\n" + textIntput.substring(prefixLength + 1, prefixLength + 70)
     );
   });
 
@@ -93,8 +92,8 @@ describe("Sorter", () => {
     );
 
     const textIntput = tcInterfaceWithCapitalLetter;
-    const { nodes } = parser.parseInterface(filePath, textIntput);
-    const sorted = sorter2.sortInterfaceElements(nodes);
+    const { nodes } = parser.parseTypeNodes(filePath, textIntput);
+    const sorted = sorter2.sortGenericTypeElements(nodes);
 
     expect(sorted.length).toBe(1);
     expect(sorted[0].rangeToRemove).toEqual({ pos: 23, end: 82 });
@@ -111,8 +110,8 @@ describe("Sorter", () => {
     );
 
     const textIntput = tcInterfaceWithCapitalLetter;
-    const { nodes } = parser.parseInterface(filePath, textIntput);
-    const sorted = sorter2.sortInterfaceElements(nodes);
+    const { nodes } = parser.parseTypeNodes(filePath, textIntput);
+    const sorted = sorter2.sortGenericTypeElements(nodes);
 
     expect(sorted.length).toBe(1);
     expect(sorted[0].rangeToRemove).toEqual({ pos: 23, end: 82 });
@@ -134,8 +133,8 @@ describe("Sorter", () => {
       );
 
       const textIntput = tcInterfaceWithOptionalProperty;
-      const { nodes } = parser.parseInterface(filePath, textIntput);
-      const sorted = sorter2.sortInterfaceElements(nodes);
+      const { nodes } = parser.parseTypeNodes(filePath, textIntput);
+      const sorted = sorter2.sortGenericTypeElements(nodes);
 
       expect(sorted.length).toBe(1);
       expect(sorted[0].rangeToRemove).toEqual({ pos: 23, end: 73 });
@@ -156,8 +155,8 @@ describe("Sorter", () => {
       );
 
       const textIntput = tcInterfaceWithMultipleOptionalProperty;
-      const { nodes } = parser.parseInterface(filePath, textIntput);
-      const sorted = sorter2.sortInterfaceElements(nodes);
+      const { nodes } = parser.parseTypeNodes(filePath, textIntput);
+      const sorted = sorter2.sortGenericTypeElements(nodes);
 
       expect(sorted.length).toBe(1);
       expect(sorted[0].rangeToRemove).toEqual({ pos: 23, end: 125 });
@@ -178,8 +177,8 @@ describe("Sorter", () => {
       );
 
       const textIntput = tcInterfaceWithMultipleOptionalProperty;
-      const { nodes } = parser.parseInterface(filePath, textIntput);
-      const sorted = sorter2.sortInterfaceElements(nodes);
+      const { nodes } = parser.parseTypeNodes(filePath, textIntput);
+      const sorted = sorter2.sortGenericTypeElements(nodes);
 
       expect(sorted.length).toBe(1);
       expect(sorted[0].rangeToRemove).toEqual({ pos: 23, end: 125 });